Fully refactor firmware out of soccer
