+++ soccer/radio/USBRadio.cpp
@@ -4,15 +4,14 @@
 #include <QMutexLocker>
 
 #include <Utils.hpp>
-#include "../firmware/common2015/drivers/cc1201/ti/defines.hpp"
+#include "../common/firmware/common2015/drivers/cc1201/ti/defines.hpp"
 #include "USBRadio.hpp"
 
 // Include this file for base station usb vendor/product ids
-#include "../firmware/base2015/usb-interface.hpp"
+#include "../common/firmware/base2015/usb-interface.hpp"
 
 // included for kicer status enum
-#include "../firmware/robot2011/cpu/status.h"
-
+#include "../common/firmware/robot2015/cpu/status.h"
 
 using namespace std;
 using namespace Packet;
